Fixed some integer type errors/warnings

+++ MathIndex.h
@@ -33,7 +33,7 @@ public:
 	~MathIndex();
 	
 	int NumberOfImages();
-	unsigned char* GetImage(string filename, int* size);
+	unsigned char* GetImage(string filename, size_t* size);
 	
 private:
 	string	_dataFileName;
